Move markdownwithpreview component to rich-text-editor folder
