+++ openmetadata-ui/src/main/resources/ui/src/components/AddDataQualityTest/Forms/TableTestForm.tsx
@@ -27,7 +27,7 @@ import {
   requiredField,
 } from '../../../utils/CommonUtils';
 import { Button } from '../../buttons/Button/Button';
-import MarkdownWithPreview from '../../common/editor/MarkdownWithPreview';
+import MarkdownWithPreview from '../../common/rich-text-editor/MarkdownWithPreview';
 
 type Props = {
   data?: TableTest;
