+++ openmetadata-ui/src/main/resources/ui/src/pages/teams/Form.tsx
@@ -19,7 +19,7 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import MarkdownWithPreview from '../../components/common/editor/MarkdownWithPreview';
+import MarkdownWithPreview from '../../components/common/rich-text-editor/MarkdownWithPreview';
 import { errorMsg } from '../../utils/CommonUtils';
 
 type FormProp = {
